Move EventType and Algorithm outside
